server/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql==0.0.3
mysqlclient==2.1.1
numpy==2.0.2
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
XlsxWriter==3.2.2
zipp==3.21.0
mysql-connector-python==8.0.29
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.3
flask_login 